MDTeamGPT_6_NewArc_1/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
flake8==6.1.0          # Code linter
mypy==1.8.0            # Static type checker
pytest==7.4.4          # Testing framework
numpy==1.26.4
pandas==1.5.3
scikit-learn==1.6.1
pyarrow==14.0.2
sentence-transformers==2.2.2
datasets==2.14.6
faiss-cpu>=1.7.2
sentence-transformers>=2.2.2
rank_bm25>=0.2.1

